requirements.txt
gunicorn==23.0.0
Flask==3.0.0
pymongo==4.7.0
bcrypt==4.1.2
openai==1.57.0
plotly==5.21.0
pandas==2.2.2
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.1
celery==5.3.4
redis==5.0.1
beautifulsoup4==4.12.3
nltk==3.8.1
pytz==2024.1
spacy==3.7.2
hdbscan==0.8.33
scikit-learn==1.4.0
umap-learn==0.5.5
numpy==1.26.4